flytter filter-predikatene
